requirements.txt
absl-py==0.7.0
astor==0.7.1
boto==2.49.0
boto3==1.9.96
botocore==1.12.96
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configobj==5.0.6
docutils==0.14
fasttext==0.8.3
Flask==1.0.2
future==0.17.1
gast==0.2.2
gensim==3.7.1
grpcio==1.18.0
h5py==2.9.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonpatch==1.21
jsonpointer==2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.16.1
oauth==1.0.1
pandas==0.24.1
pathlib==1.0.1
pdfminer.six==20181108
prettytable==0.7.2
protobuf==3.6.1
pycryptodome==3.7.3
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==4.2b1
requests==2.20.0
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.11.0
sklearn==0.0
smart-open==1.8.0
sortedcontainers==2.1.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.23 
uWSGI==2.0.17.1
waitress==1.1.0
Werkzeug==0.14.1
